requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
PyMuPDF==1.23.7
llama-cpp-python==0.2.72